vp test: create setup for esm preview
