+++ test/e2e/specs/linksConsoleErrorsEsmPage.spec.ts
@@ -1,16 +1,10 @@
-import { ConsoleMessage, expect, Page } from '@playwright/test';
+import { ConsoleMessage, expect } from '@playwright/test';
 import { vpTest } from '../fixtures/vpTest';
 import { ESM_LINKS } from '../testData/esmPageLinksData';
 import { waitForPageToLoadWithTimeout } from '../src/helpers/waitForPageToLoadWithTimeout';
 import { validatePageErrors } from '../src/helpers/validatePageErrors';
-import { ExampleLinkType } from '../types/exampleLinkType';
 import { ExampleLinkName } from '../testData/ExampleLinkNames';
-
-const EDGE_ESM_URL = 'https://cld-vp-esm-pages.netlify.app/';
-// On PR level it will use the preview deploy URL and locally it will use the latest EDGE.
-const ESM_URL = process.env.PREVIEW_URL ?? EDGE_ESM_URL;
-// Flag to indicate if the deploy preview URL is ready
-let isPreviewUrlLoaded = false;
+import { ESM_URL } from '../testData/esmUrl';
 
 /**
  * Console error test generated by LINKS object array data.
@@ -18,10 +12,6 @@ let isPreviewUrlLoaded = false;
 for (const link of ESM_LINKS) {
     vpTest(`Test console errors on link ${link.name}`, async ({ page, consoleErrors, pomPages }) => {
         vpTest.skip(link.name === 'Adaptive Streaming', 'Flaky on CI');
-        //Wait for deploy URL to be available if PREVIEW_URL is set, and it is not available yet
-        if (process.env.PREVIEW_URL && !isPreviewUrlLoaded) {
-            await waitForDeployPreviewUrl(link, page);
-        }
         await page.goto(ESM_URL);
         await pomPages.mainPage.clickLinkByName(link.name);
         await waitForPageToLoadWithTimeout(page, 5000);
@@ -57,16 +47,4 @@ function handleCommonEsmBrowsersErrors(linkName: ExampleLinkName, consoleErrors:
         default:
             validatePageErrors(consoleErrors, [], ['the server responded with a status of 404']);
     }
-}
-
-/**
- * Waits for a deploy preview URL to become available by making repeated requests and check that link is visible.
- */
-async function waitForDeployPreviewUrl(link: ExampleLinkType, page: Page): Promise<void> {
-    await expect(async () => {
-        await page.goto(process.env.PREVIEW_URL);
-        const linkLocator = page.getByRole('link', { name: link.name, exact: true });
-        await expect(linkLocator).toBeVisible({ timeout: 10000 });
-        isPreviewUrlLoaded = true;
-    }).toPass({ intervals: [1_000], timeout: 120000 });
 }
