+++ test/e2e/testData/esmUrl.ts
@@ -1,0 +1,2 @@
+// On PR level it will use the preview deploy URL and locally it will use the latest EDGE.
+export const ESM_URL = process.env.PLAYWRIGHT_TEST_BASE_URL ?? 'https://cld-vp-esm-pages.netlify.app/';
